设备打包发布按钮及进度条展示

+++ engine/device-generator/src/main/java/lowcode/device/generator/core/DeviceGenerator.java
@@ -11,6 +11,7 @@ import org.springframework.boot.configurationprocessor.json.JSONException;
 import org.springframework.boot.configurationprocessor.json.JSONObject;
 
 import java.io.File;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
@@ -75,17 +76,22 @@ public class DeviceGenerator {
         generateServiceJson();
     }
 
-    public void buildAndPackage(){
+    public void buildAndPackage() throws Exception {
         projectGenerator.buildAndPackage(new File(parentPath+deviceType.toLowerCase()));
+    }
 
+    public void copyJarFile() throws Exception{
         // 生成的jar包拷贝到工作目录
-        try {
-            File source = new File(parentPath+deviceType.toLowerCase()+"/target/"+deviceType.toLowerCase()+"-"+version+".jar");
-            File dest = new File(definitionPath + "generate");
-            FileUtil.copyFile(source, dest);
-        }catch (Exception e) {
-            e.printStackTrace();
+        File source = new File(parentPath+deviceType.toLowerCase()+"/target/"+deviceType.toLowerCase()+"-"+version+".jar");
+        if (!source.exists()){
+            throw new RuntimeException("拷贝失败，jar包不存在");
         }
+
+        File dest = new File(definitionPath + "generate");
+        if (!dest.exists()){
+            dest.mkdirs();
+        }
+        FileUtil.copyFile(source, dest);
     }
 
     public void generateEventJson(String jsonPath) {
